Move components into subfolders
