+++ client/src/components/Admin/Home.js
@@ -1,6 +1,6 @@
 import React from 'react';
-import AdminIssueContainer from '../containers/AdminIssueContainer';
-import AdminPanelAlternativesContainer from '../containers/AdminPanelAlternativesContainer';
+import AdminIssueContainer from '../../containers/AdminIssueContainer';
+import AdminPanelAlternativesContainer from '../../containers/AdminPanelAlternativesContainer';
 
 
 const AdminHome = () => (
