--- client/src/components/AdminPanelAlternatives.js
@@ -1,6 +1,0 @@
-import React from 'react';
-import Alternatives from './Alternatives';
-
-export default ({ alternatives }) => (
-  <Alternatives alternatives={alternatives} />
-);
